refactor(ui): move the text input into the design system
