+++ crates/input/Cargo.toml
@@ -6,5 +6,4 @@ license.workspace = true
 
 [dependencies]
 gpui.workspace = true
-i18n.workspace = true
 ui.workspace = true
