+++ crates/ui/src/form.rs
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+use gpui::actions;
+
+actions!(form, [Submit]);
+
+pub const FORM_CONTEXT: &str = "Form";
