+++ crates/ui/src/lib.rs
@@ -6,9 +6,11 @@ mod card;
 mod controls;
 mod explicit;
 mod filters;
+mod form;
 mod grid;
 mod info_card;
 mod inline_links;
+mod input;
 mod label;
 mod layout;
 mod menu;
@@ -33,6 +35,7 @@ pub use card::Card;
 pub use controls::WindowControls;
 pub use explicit::ExplicitBadge;
 pub use filters::{FlagAxis, RangeAxis, RangeScrubber, RangeState, SortAxis, Unit};
+pub use form::{FORM_CONTEXT, Submit};
 pub use grid::{
     Cell, ColumnSpec, Deselect, GRID_CONTEXT, Grid, GridDelegate, GridEvent, GridSource, GridState,
     Layout, ROW_GROUP, SelectNext, SelectPrevious, Sort, Sorting, Table, Toggle, Viewport, Width,
@@ -40,6 +43,11 @@ pub use grid::{
 };
 pub use info_card::{Fact, InfoCard};
 pub use inline_links::{InlineLink, InlineLinks};
+pub use input::{
+    Backspace, BackspaceWord, Copy, Cut, Delete, DeleteWord, Dismiss, End, Home, INPUT_CONTEXT,
+    Input, Left, Paste, Right, SelectAll, SelectEnd, SelectHome, SelectLeft, SelectRight,
+    SelectWordLeft, SelectWordRight, Space, WordLeft, WordRight,
+};
 pub use label::{eyebrow, heading, upper};
 pub use layout::{ALWAYS, MIN_CONTENT, ROOMY, Room, SNUG, VAST, WIDE};
 pub use menu::{Menu, MenuItem, SubmenuState};
